+++ packages/decap-cms-core/src/components/Collection/Collection.js
@@ -27,16 +27,23 @@ import {
 import { selectCanCreateNewEntry } from '../../reducers';
 
 const CollectionContainer = styled.div`
-  margin: ${lengths.pageMargin};
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin: ${lengths.pageMarginMobile};
+  @media (min-width: 500px) {
+    margin: ${lengths.pageMargin};
+  }
 `;
 
 const CollectionMain = styled.main`
-  padding-left: 280px;
+  @media (min-width: 800px) {
+    padding-left: 280px;
+  }
 `;
 
 const SearchResultContainer = styled.div`
   ${components.cardTop};
-  margin-bottom: 22px;
 `;
 
 const SearchResultHeading = styled.h1`