+++ packages/decap-cms-core/src/components/Collection/CollectionTop.js
@@ -8,11 +8,12 @@ import { components, buttons, shadows } from 'decap-cms-ui-default';
 
 const CollectionTopContainer = styled.div`
   ${components.cardTop};
-  margin-bottom: 22px;
 `;
 
 const CollectionTopRow = styled.div`
   display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
   align-items: center;
   justify-content: space-between;
 `;
@@ -26,13 +27,13 @@ const CollectionTopNewButton = styled(Link)`
   ${shadows.dropDeep};
   ${buttons.default};
   ${buttons.gray};
+  white-space: nowrap;
 
   padding: 0 30px;
 `;
 
 const CollectionTopDescription = styled.p`
   ${components.cardTopDescription};
-  margin-bottom: 0;
 `;
 
 function getCollectionProps(collection) {
@@ -58,7 +59,13 @@ function CollectionTop({ collection, newEntryUrl, t }) {
       <CollectionTopRow>
         <CollectionTopHeading>{collectionLabel}</CollectionTopHeading>
         {newEntryUrl ? (
-          <CollectionTopNewButton to={newEntryUrl}>
+          <CollectionTopNewButton
+            to={newEntryUrl}
+            dir="auto"
+            aria-label={t('collection.collectionTop.newButtonAriaLabel', {
+              collectionLabel: collectionLabelSingular || collectionLabel,
+            })}
+          >
             {t('collection.collectionTop.newButton', {
               collectionLabel: collectionLabelSingular || collectionLabel,
             })}