+++ packages/decap-cms-core/src/components/Workflow/WorkflowList.js
@@ -15,47 +15,57 @@ import { selectEntryCollectionTitle } from '../../reducers/collections';
 const WorkflowListContainer = styled.div`
   min-height: 60%;
   display: grid;
-  grid-template-columns: 33.3% 33.3% 33.3%;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 8px;
+  @media (min-width: 500px) {
+    gap: 14px;
+  }
+  @media (min-width: 800px) {
+    gap: 40px;
+  }
 `;
 
 const WorkflowListContainerOpenAuthoring = styled.div`
   min-height: 60%;
   display: grid;
-  grid-template-columns: 50% 50% 0%;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 8px;
+  @media (min-width: 500px) {
+    gap: 14px;
+  }
+  @media (min-width: 800px) {
+    gap: 40px;
+  }
 `;
 
 const styles = {
   columnPosition: idx =>
-    (idx === 0 &&
-      css`
-        margin-left: 0;
-      `) ||
-    (idx === 2 &&
-      css`
-        margin-right: 0;
-      `) ||
+    idx > 0 &&
     css`
-      &:before,
-      &:after {
+      &:before {
         content: '';
         display: block;
         position: absolute;
         width: 2px;
         height: 80%;
-        top: 76px;
+        top: 36px;
         background-color: ${colors.textFieldBorder};
+        @media (min-width: 800px) {
+          top: 76px;
+        }
       }
 
       &:before {
-        left: -23px;
-      }
-
-      &:after {
-        right: -23px;
+        left: -7px;
+        @media (min-width: 500px) {
+          left: -10px;
+        }
+        @media (min-width: 800px) {
+          left: -23px;
+        }
       }
     `,
   column: css`
-    margin: 0 20px;
     transition: background-color 0.5s ease;
     border: 2px dashed transparent;
     border-radius: 4px;
@@ -68,21 +78,21 @@ const styles = {
   hiddenColumn: css`
     display: none;
   `,
-  hiddenRightBorder: css`
-    &:not(:first-child):not(:last-child) {
-      &:after {
-        display: none;
-      }
-    }
-  `,
 };
 
 const ColumnHeader = styled.h2`
-  font-size: 20px;
+  font-size: 16px;
   font-weight: normal;
-  padding: 4px 14px;
+  padding: 2px 6px;
   border-radius: ${lengths.borderRadius};
-  margin-bottom: 28px;
+  white-space: nowrap;
+  margin-bottom: 6px;
+
+  @media (min-width: 800px) {
+    font-size: 20px;
+    padding: 4px 12px;
+    margin-bottom: 28px;
+  }
 
   ${props =>
     props.name === 'draft' &&
@@ -107,11 +117,21 @@ const ColumnHeader = styled.h2`
 `;
 
 const ColumnCount = styled.p`
-  font-size: 13px;
+  font-size: 11px;
   font-weight: 500;
   color: ${colors.text};
   text-transform: uppercase;
   margin-bottom: 6px;
+  padding-inline: 6px;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  min-width: 0;
+
+  @media (min-width: 800px) {
+    font-size: 13px;
+    padding-inline: 12px;
+  }
 `;
 
 // This is a namespace so that we can only drop these elements on a DropTarget with the same
@@ -188,7 +208,6 @@ class WorkflowList extends React.Component {
                     styles.columnPosition(idx),
                     isHovered && styles.columnHovered,
                     isOpenAuthoring && currColumn === 'pending_publish' && styles.hiddenColumn,
-                    isOpenAuthoring && currColumn === 'pending_review' && styles.hiddenRightBorder,
                   ]}
                 >
                   <ColumnHeader name={currColumn}>