+++ packages/decap-cms-core/src/types/redux.ts
@@ -11,6 +11,7 @@ import type { Search } from '../reducers/search';
 import type { GlobalUI } from '../reducers/globalUI';
 import type { NotificationsState } from '../reducers/notifications';
 import type { formatExtensions } from '../formats/formats';
+import type { Note } from 'decap-cms-lib-util';
 
 export type CmsBackendType =
   | 'azure'
@@ -72,7 +73,7 @@ export interface CmsFieldBase {
   label?: string;
   required?: boolean;
   hint?: string;
-  pattern?: [string, string];
+  pattern?: [string | RegExp, string];
   i18n?: boolean | 'translate' | 'duplicate' | 'none';
   media_folder?: string;
   public_folder?: string;
@@ -312,17 +313,38 @@ export interface CmsCollectionFile {
   public_folder?: string;
 }
 
+export type NoteMap = StaticallyTypedRecord<Note>;
+
+export type Notes = List<NoteMap>;
+
+export interface LoadNotesPayload {
+  notes: Note[];
+}
+
+export interface AddNotePayload {
+  note: Note;
+}
+
+export interface UpdateNotePayload {
+  id: string;
+  updates: Partial<Note>;
+}
+
+export interface DeleteNotePayload {
+  id: string;
+}
+
 export interface ViewFilter {
   label: string;
   field: string;
-  pattern: string;
+  pattern: string | boolean;
   id: string;
 }
 
 export interface ViewGroup {
   label: string;
   field: string;
-  pattern: string;
+  pattern?: string;
   id: string;
 }
 
@@ -349,6 +371,7 @@ export interface CmsCollection {
   delete?: boolean;
   editor?: {
     preview?: boolean;
+    notes?: boolean;
     visualEditing?: boolean;
   };
   publish?: boolean;
@@ -455,6 +478,7 @@ export interface CmsConfig {
   local_backend?: boolean | CmsLocalBackend;
   editor?: {
     preview?: boolean;
+    notes?: boolean;
   };
   error: string | undefined;
   isFetching: boolean;
@@ -587,6 +611,7 @@ export type EntryDraft = StaticallyTypedRecord<{
   entry: Entry;
   fieldsErrors: FieldsErrors;
   fieldsMetaData?: Map<string, Map<string, string>>;
+  notes: Notes;
 }>;
 
 export type EntryField = StaticallyTypedRecord<{
@@ -611,6 +636,12 @@ export type FilterRule = StaticallyTypedRecord<{
   field: string;
 }>;
 
+type CollectionEditor = StaticallyTypedRecord<{
+  preview?: boolean;
+  notes?: boolean;
+  visualEditing?: boolean;
+}>;
+
 export type CollectionFile = StaticallyTypedRecord<{
   file: string;
   name: string;
@@ -620,6 +651,7 @@ export type CollectionFile = StaticallyTypedRecord<{
   public_folder?: string;
   preview_path?: string;
   preview_path_date_field?: string;
+  editor?: CollectionEditor;
 }>;
 
 export type CollectionFiles = List<CollectionFile>;
@@ -663,6 +695,7 @@ type CollectionObject = {
   frontmatter_delimiter?: List<string> | string | [string, string];
   create?: boolean;
   delete?: boolean;
+  editor?: CollectionEditor;
   identifier_field?: string;
   path?: string;
   slug?: string;